move federation config to graphql/federation pkg
