+++ pkg/graphql/federation/federation_config_test.go
@@ -1,0 +1,2 @@
+package federation
+
